Stress BC: A Tensor Solver Boundary Condition
The new boundary condition is tau dot n = f n, where tau is the stress
tensor and n is normal vector for the domain face.  Here, f is zero, unless
it's provided by the user with a call to MLTensorOp::setInhomogStressBC.

This is the first pass.  It's still not clear how to handle properly some of
the terms.  More specifically, how to compute du/dx and dv/dx on y-faces
of the cells next the x-face of the domain?